Move to build system in appium-gulp-plugins
