--- apps.json
@@ -1,4 +1,0 @@
-[
-  "build/Release-iphoneos/TestApp-iphoneos.app",
-  "build/Release-iphonesimulator/TestApp-iphonesimulator.app"
-]
