+++ index.js
@@ -1,5 +1,19 @@
-import { relative, absolute, appList, install, installRealDevice } from './install-npm';
+'use strict';
+
+const path = require('path');
 
 
-export default appList;
-export { relative, absolute, appList, install, installRealDevice };
+const relative = {
+  iphoneos: 'build/Release-iphoneos/TestApp-iphoneos.app',
+  iphonesimulator: 'build/Release-iphonesimulator/TestApp-iphonesimulator.app'
+};
+
+const absolute = {
+  iphoneos: path.resolve('build', 'Release-iphoneos', 'TestApp-iphoneos.app'),
+  iphonesimulator: path.resolve('build', 'Release-iphonesimulator', 'TestApp-iphonesimulator.app')
+};
+
+module.exports = {
+  relative,
+  absolute,
+};
