--- lib/logger.js
@@ -1,5 +1,0 @@
-import { logger } from 'appium-support';
-
-let log = logger.getLogger('iOSTestApp');
-
-export default log;
