+++ package.json
@@ -22,26 +22,26 @@
   "os": [
     "darwin"
   ],
-  "main": "build-js/index.js",
+  "main": "index.js",
   "bin": {},
   "directories": {
     "lib": "lib"
   },
-  "dependencies": {
-    "@babel/runtime": "^7.0.0",
-    "appium-support": "^2.5.0",
-    "appium-xcode": "^3.6.0",
-    "asyncbox": "^2.3.1",
-    "fancy-log": "^1.3.2",
-    "source-map-support": "^0.5.6",
-    "teen_process": "^1.5.1"
-  },
+  "files": [
+    "index.js",
+    "build-js",
+    "lib",
+    "Test App 2",
+    "TestApp.xcodeproj"
+  ],
+  "dependencies": {},
   "scripts": {
     "clean": "rm -rf node_modules && rm -f package-lock.json && npm install",
-    "prepublish": "./node_modules/.bin/gulp prepublish",
-    "install": "node install-npm.js",
-    "test": "./node_modules/.bin/gulp once",
-    "watch": "./node_modules/.bin/gulp",
+    "prepare": "gulp prepublish",
+    "build": "gulp ios-apps:install",
+    "install": "gulp ios-apps:install",
+    "test": "gulp once",
+    "watch": "gulp watch",
     "lint": "gulp lint",
     "lint:fix": "gulp lint --fix",
     "precommit-msg": "echo 'Pre-commit checks...' && exit 0",
@@ -53,7 +53,7 @@
   ],
   "devDependencies": {
     "ajv": "^6.5.3",
-    "appium-gulp-plugins": "^3.0.0",
+    "appium-gulp-plugins": "^3.7.0",
     "babel-eslint": "^10.0.1",
     "eslint": "^5.2.0",
     "eslint-config-appium": "^4.0.1",
